copy_file: fix progress bar for files larger than 2G
Files larger than 2G cause the benign bug, that the progress bar is no
longer accurate. Use loff_t to fix this.

Note that printed % HASHES_PER_LINE in show_progress implies a 64-bit
division. A previous commit changed the constant divisor to a power of
two, so a division shouldn't be generated. The original multiplication
and modulo operations are left as is for clarity.

+++ include/progress.h
@@ -2,15 +2,17 @@
 #ifndef __PROGRSS_H
 #define __PROGRSS_H
 
+#include <linux/types.h>
+
 /* Initialize a progress bar. If max > 0 a one line progress
  * bar is printed where 'max' corresponds to 100%. If max == 0
  * a multi line progress bar is printed.
  */
-void init_progression_bar(int max);
+void init_progression_bar(loff_t max);
 
 /* update a progress bar to a new value. If now < 0 then a
  * spinner is printed.
  */
-void show_progress(int now);
+void show_progress(loff_t now);
 
 #endif /*  __PROGRSS_H */
